+++ src/test/groovy/edu/ie3/simona/model/participant/ChpModelTest.groovy
@@ -88,8 +88,8 @@ class ChpModelTest extends Specification {
         thermalStorage)
   }
 
-  static def buildChpData(ChpState chpState, Double heatDemand) {
-    return new ChpModel.ChpData(chpState, getQuantity(heatDemand, KILOWATTHOUR), 7200)
+  static def buildChpRelevantData(ChpState chpState, Double heatDemand) {
+    return new ChpModel.ChpRelevantData(chpState, getQuantity(heatDemand, KILOWATTHOUR), 7200)
   }
 
   static def buildThermalStorage(CylindricalStorageInput storageInput, Double storageLvl) {
@@ -101,7 +101,7 @@ class ChpModelTest extends Specification {
   @Unroll
   def "Check active power after calculating next state with #chpState and heat demand #heatDemand kWh:"() {
     given:
-    def chpData = buildChpData(chpState, heatDemand)
+    def chpData = buildChpRelevantData(chpState, heatDemand)
     def thermalStorage = buildThermalStorage(storageInput, storageLvl)
     def chpModel = buildChpModel(thermalStorage)
 
@@ -128,7 +128,7 @@ class ChpModelTest extends Specification {
   @Unroll
   def "Check total energy after calculating next state with #chpState and heat demand #heatDemand kWh:"() {
     given:
-    def chpData = buildChpData(chpState, heatDemand)
+    def chpData = buildChpRelevantData(chpState, heatDemand)
     def thermalStorage = buildThermalStorage(storageInput, storageLvl)
     def chpModel = buildChpModel(thermalStorage)
 
@@ -156,7 +156,7 @@ class ChpModelTest extends Specification {
 
   def "Check storage level after calculating next state with #chpState and heat demand #heatDemand kWh:"() {
     given:
-    def chpData = buildChpData(chpState, heatDemand)
+    def chpData = buildChpRelevantData(chpState, heatDemand)
     def thermalStorage = buildThermalStorage(storageInput, storageLvl)
     def chpModel = buildChpModel(thermalStorage)
 
@@ -184,7 +184,7 @@ class ChpModelTest extends Specification {
 
   def "Check time tick and running status after calculating next state with #chpState and heat demand #heatDemand kWh:"() {
     given:
-    def chpData = buildChpData(chpState, heatDemand)
+    def chpData = buildChpRelevantData(chpState, heatDemand)
     def thermalStorage = buildThermalStorage(storageInput, storageLvl)
     def chpModel = buildChpModel(thermalStorage)